+++ tests/integration-test-config.php
@@ -230,7 +230,6 @@ return [
 	 * will be emitted by Phan.
 	 */
 	'whitelist_issue_types' => [
-		'SecurityCheckMulti',
 		'SecurityCheck-XSS',
 		'SecurityCheck-SQLInjection',
 		'SecurityCheck-ShellInjection',